Gold is now a float, added end game check

+++ GameSource/Scripts/Player/MultiplayerPlayerClass.cs
@@ -47,7 +47,7 @@ public partial class MultiplayerPlayerClass : Node
 		var packet = ShopItemBuy.CreateFromData(data);
 		uiCommunicator.shopCards.RemoveCard(uiCommunicator.shopCards.GetChild(packet.CardIndex) as Node3D, false);
 		PlayerClass.Gold = packet.GoldAmount;
-		_playerHud.UpdateGoldAmount(PlayerClass.Gold);
+		_playerHud.UpdateGoldAmount((int)PlayerClass.Gold);
 	}
 
 	private void SetGoldAmount(byte[] data)
@@ -56,7 +56,7 @@ public partial class MultiplayerPlayerClass : Node
 		PlayerClass.Gold += packet.GoldAmount;
 		PlayerClass.Points -= packet.PointAmount;
 		_playerHud.UpdatePointsAmount(PlayerClass.Points);
-		_playerHud.UpdateGoldAmount(PlayerClass.Gold);
+		_playerHud.UpdateGoldAmount((int)PlayerClass.Gold);
 	}
 
 	private void SetTargetPosition(byte[] data)
